# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SendaSnap API",
        "description": "Vehicle photo upload and workflow management system",
        "contact": {
            "email": "admin@sendasnap.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://sendasnap-backend-laravel.test",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/vehicles/search": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Search vehicles (senda.us avis_03_oct table)",
                "description": "Searches external vehicle database using header inputs",
                "operationId": "e62ddf990d7c6b316edddc67b5a7fe0d",
                "parameters": [
                    {
                        "name": "search_type",
                        "in": "query",
                        "description": "Type of identifier (vehicle_id or veh_chassis_number)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "vehicle_id",
                                "veh_chassis_number"
                            ]
                        }
                    },
                    {
                        "name": "search_query",
                        "in": "query",
                        "description": "Value of the identifier to search, example 251144",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search completed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}